added lazystore and hadjustment computation in size-allocate
I added the new lazystore which provides the GtkTreeModel interface to
access data without actually storing the data. This replaces the GtkListStore
which stores all data internally. For a large number of columns and rows
this is prohibitive.

In addition I added the hadjustment computation in a size-allocate method
based on the number of columns in the store and the number of columns in
the model.